Normalize keyval for position before use
Korean input method depends on the key position on the keyboard, not the key
value. So we need to change the keyval to some generic value which can
represent the position on the keyboard. I choose the US qwerty layout as the
generic value. Before passing the keyval to libhangul, we change the keyval
as if the keyboard is US qwerty keyboard. Then, libhangul can handle the
keyval as the right position.
This will fix the problem with non-qwerty keyboard users.