issue: #1210 hangul_keyboard_list should be moved
hangul_keyboard_list is architecture dependent file, so it should be
installed in ${pkglibdir} or its subdir. I made it installed in
${pkglibdir}/setup.

config.py will be created by Makefile not configure. Variable ${pkglibdir}
is suitable for Make, but python will not expand it correctly. It's better
for Make to generate config.py so that Make expand the variables.