Fix wrong space order problem
With IBUS_ENABLE_SYNC_MODE=1 option,
'rksk ' is translated '가 나'.
It's caused by limitations of IBus implementation.
IBus can not send a result of other function inside a function
in sync mode. The results of other functions will be delayed until
the current function ends.
This makes ibus-hangul do wrong.

We solve this problem by forwarding key event.
It may be a hack, but we can not help it.
That's the limitation of the framework.

https://github.com/choehwanjin/ibus-hangul/issues/42