Use bind_textdomain_codeset() from gettext.
bind_textdomain_codeset() is part of the locale module on linux, but not
on openbsd.  It is always included in the gettext module though, which
is in fact the correct module to use.
From https://docs.python.org/2/library/locale.html:

The locale module exposes the C library’s gettext interface on systems
that provide this interface. It consists of the functions gettext(),
dgettext(), dcgettext(), textdomain(), bindtextdomain(), and
bind_textdomain_codeset(). These are similar to the same functions in
the gettext module, but use the C library’s binary format for message
catalogs, and the C library’s search algorithms for locating message
catalogs.