setup: Set GLib app name explicitly
ibus-setup-hangul program is called as "python .../main.py" command line so
the argv[0] "main.py" was used as the name as default. This name appeared in
the windows list. (See https://stackoverflow.com/questions/11854142)

So the app name is explicitly set using GLib.set_prgname() and
GLib.set_application_name().