Reparent IBusHangulEngine from IBusEngineSimple to handle Compose key.
The compose key fails to be interpreted in the latin mode of
Ibus-Hangul which is annoying because when you set this feature, you
expect it to work with every input. IBusEngineSimple handles key
composition already so the solution is to reparent IBusHangulEngine and
to delegate to the parent's process_key_event() when in latin mode.